Change sensorvalue type in 'double'

+++ src/masschange/dataproducts/implementations/gracefo/primary/lhk1a.py
@@ -31,7 +31,7 @@ class GraceFOLhk1ADataProduct(TimeSeriesDataProduct):
             qualflg VARCHAR(8) not null,
             
             sensortype CHAR not null,
-            sensorvalue bigint not null,
+            sensorvalue double precision not null,
             sensorname VARCHAR(1000) not null,
             unit VARCHAR(4),
